+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Changelog
 **Next release**
 
 * Added restriction *Srv_startActivityAsCaller* ([issue](/../../issues/1757))
+* Added restriction *Srv_getAccountsForPackage*
+* Added account type parameter for *Srv_getAccounts* and *Srv_getAccountsAsUser*
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)